Move and rename schedule sale block to the blocks folder
