+++ packages/js/product-editor/src/blocks/schedule-sale/editor.scss
@@ -1,4 +1,4 @@
-.wp-block-woocommerce-product-schedule-sale-pricing {
+.wp-block-woocommerce-product-schedule-sale-fields {
 	.components-toggle-control__label {
 		display: flex;
 		align-items: center;
