+++ packages/js/product-editor/src/style.scss
@@ -1,3 +1,4 @@
+@import 'blocks/schedule-sale/editor.scss';
 @import 'components/editor/style.scss';
 @import 'components/product-section-layout/style.scss';
 @import 'components/edit-product-link-modal/style.scss';
@@ -13,4 +14,3 @@
 @import 'components/details-summary-block/style.scss';
 @import 'components/product-mvp-ces-footer/style.scss';
 @import 'components/product-mvp-feedback-modal/style.scss';
-@import 'components/schedule-sale-pricing-block/editor.scss';
